+++ CHANGELOG.md
@@ -4,5 +4,6 @@
     * Fix: Migration alert message.
     * Fix: Updated Jigoshop logotypes.
     * Fix: Admin Order and Product list search.
+    * Fix: Fatal error on admin order item variable template.
 * 2.0.0 - 2016.10.27:
     * Full rewrite of core.