+++ CHANGELOG.md
@@ -6,5 +6,6 @@
     * Fix: Admin Order and Product list search.
     * Fix: Fatal error on admin order item variable template.
     * Fix: Paypal response handling.
+    * Fix: Duplicatd foregin keys.
 * 2.0.0 - 2016.10.27:
     * Full rewrite of core.