+++ CHANGELOG.md
@@ -8,5 +8,6 @@
     * Fix: Paypal response handling.
     * Fix: Duplicatd foregin keys.
     * Fix: Show add to cart button for free products.
+    * Fix: Change order status if payment is not required.
 * 2.0.0 - 2016.10.27:
     * Full rewrite of core.