+++ CHANGELOG.md
@@ -9,5 +9,6 @@
     * Fix: Duplicatd foregin keys.
     * Fix: Show add to cart button for free products.
     * Fix: Change order status if payment is not required.
+    * Fix: Properly clear cart after checkout.
 * 2.0.0 - 2016.10.27:
     * Full rewrite of core.